+++ docs/changelog.md
@@ -1,5 +1,6 @@
 # Latest 
 - Refactor to separate omnibar and search
+- Add CTRL+P for omnibar focus
 
 # Version 3.5.0: More useful search 
 - Fix minor focus bug